cod atual 18/04

+++ src/domain/Price.java
@@ -1,24 +1,23 @@
 package domain;
 
-import java.math.BigDecimal;
+import java.lang.Float;
 import java.util.Date;
 import java.util.UUID;
 
 public class Price implements EntityInterface{
     private UUID uuid;
-    private BigDecimal price;
+    private Float price;
     private Date date;
-
 
     public Price() {
     }
 
-    public Price(BigDecimal price, Date date) {
+    public Price(Float price, Date date) {
         this.price = price;
         this.date = date;
     }
 
-    public Price(UUID uuid, BigDecimal price, Date date) {
+    public Price(UUID uuid, Float price, Date date) {
         this.uuid = uuid;
         this.price = price;
         this.date = date;
@@ -28,7 +27,6 @@ public class Price implements EntityInterface{
     public UUID getUUID() {
         return this.uuid;
     }
-
 
     @Override
     public String toString() {
